refactor: clean files, rename file Item and add code const
