+++ customer-application/src/utils/codes.ts
@@ -1,0 +1,7 @@
+export const MessageCode = {
+    OK: 200,
+    NOT_FOUND: 404,
+    SERVICE_NOT_AVAILABLE: 500
+  }
+  Object.freeze(MessageCode)
+
