+++ customer-application/src/utils/item.ts
